adds ExternalStoryWidget in ExternalStoryPage
